% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dot_env.R
\name{parse_dot_line}
\alias{parse_dot_line}
\title{Parse environment variables from lines in a .env file}
\usage{
parse_dot_line(lines)
}
\arguments{
\item{lines}{Character vector containing lines from a .env file}
}
\value{
A named list where names are environment variable names
and values are their corresponding values
}
\description{
Parse environment variables from lines in a .env file
}
